pyproject.toml
[tool.poetry]
name = "chaos-playbook-engine"
version = "0.1.0"
description = "A production-grade chaos engineering framework for AI agents"
authors = ["Alberto Martinez <albertomz@gmail.com>"]
readme = "README.md"
packages = [{include = "chaos_engine", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
google-adk = {extras = ["eval"], version = "^1.19.0"}
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
matplotlib = "^3.10.7"
seaborn = "^0.13.2"
numpy = "^2.3.5"
rich = "^14.2.0"
pyautogui = "^0.9.54"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
mypy = "^1.5.0"
black = "^23.9.0"
ruff = "^0.0.292"
isort = "^5.12.0"
ipython = "^8.15.0"
pytest-html = "^4.1.1"
coverage = "^7.12.0"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
